Organize component name
