+++ src/App.js
@@ -36,7 +36,7 @@ import AdminDashboardRejected from "./components/Admin/AdminDashboardRejected";
 import CompleteAssessment from "./components/InvestorsPortal/Dashboard/CompleteAssessment";
 import AddSpecialistStart from "./components/SpecialistPortal/SpecialistCreation/AddSpecialistStart";
 import UserAssessIdeas from "./components/MyIdea/Dashboard/UserAssessIdeas";
-import CofounderStart from './components/Creator/Co-founderStart'
+import CofounderStart from './components/Creator/CofounderStart'
 
 
 class App extends Component {
